Adding the Gauss solution class

+++ Matrix/Program.cs
@@ -7,17 +7,18 @@
             Console.WriteLine("Hello, world!");
             foreach (string arg in args) Console.WriteLine(arg);
 
-            List<List<int>> llst = new List<List<int>>()
+            List<List<short>> llst = new List<List<short>>()
             {
-                new List<int> { 1, 0, 7 },
-                new List<int> {2, 0, 6 },
-                new List<int>{3, 4, 5 },
-                new List<int>{0, 3, 0 },
-                new List<int>{9, 0, 20 }
+                new List<short> { 1, 0, 7 },
+                new List<short> {2, 0, 6 },
+                new List<short>{3, 4, 5 },
+                new List<short>{0, 3, 0 },
+                new List<short>{9, 0, 20 }
             };
 
-            Matrix<int> a = MatrixMethods.ToMatrix(llst);
+            Matrix<short> a = MatrixMethods.ToMatrix(llst);
             Console.WriteLine("Max sum: " + MatrixMethods.GetMaxGold(a));
+
             foreach (var x in MatrixMethods.GetMaxGoldPath(a))
                 Console.Write(x + "-> ");
